Implement sending email
Add content type field for attachment. Implement a cron job for sending the
pending emails periodically. If error while sending email then set status
to NOT_SENT else to SENT. This is done to ensure that an that could not be
sent will be attempted to send again next time. If this is to be avoided,
then such emails should be deleted manually.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>